move test_select_buffers_json
